# api/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: LOGIN SERVICE
# during dev, should point to your local machine
#host: localhost:3000
basePath: /login
# basePath prefixes all resource
# schemes:
#   # tip: remove http to make production-grade
#   - https
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
      For accessing the API a valid JWT token must be passed in all the queries in
      the 'Authorization' header.


      A valid JWT token is generated by the API and retourned as answer of a call
      to the route /login giving a valid user & password.


      The following syntax must be used in the 'Authorization' header :

          Bearer xxxxxx.yyyyyyy.zzzzzz
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json

paths:
  /v1/login:
    get:
      tags:
        - "Login service"
      summary: "Login service"
      description: "Login to app using credentails."
      operationId: "getWelcomeMessage"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - name: "content_type"
          in: "query"
          required: true
          description: "enter your name"
          type: "string"
      responses:
        default:
          description: "successful operation"
      x-swagger-router-controller: "controller"    

    post:
      tags:
        - "Login service"
      summary: "Login service"
      description: "Login to app using credentails."
      operationId: "userDetails"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "User credentials for authentication"
          required: true
          schema:
            $ref: '#/definitions/userCredentials'
      responses:
        default:
          description: "successfull operation"
      x-swagger-router-controller: "controller"    
        
definitions:
  userCredentials:
    type: 'object'
    properties:
      email:
        type: string
      password:
         type: string
    required:
        - email
        - password
     
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
    xml:
      name: "loginUser"
    example:
      userId: "admin@mail.com"
      password: "password"
      role: "admin"